Revert "[Phi] Migrate Adam and AdamW into Phi (#40351)" (#41712)
* Revert "[Phi] Migrate Adam and AdamW into Phi (#40351)"

This reverts commit 56cd340707d2efb7b0adc8b2e32ed38630d2dab8.

* add infermeta
