+++ paddle/phi/infermeta/multiary.h
@@ -68,55 +68,6 @@ void AdamaxInferMeta(const MetaTensor& param,
                      MetaTensor* moment_out,
                      MetaTensor* inf_norm_out);
 
-void AdamInferMeta(const MetaTensor& param,
-                   const MetaTensor& grad,
-                   const MetaTensor& learning_rate,
-                   const MetaTensor& moment1,
-                   const MetaTensor& moment2,
-                   const MetaTensor& beta1_pow,
-                   const MetaTensor& beta2_pow,
-                   paddle::optional<const MetaTensor&> master_param,
-                   paddle::optional<const MetaTensor&> skip_update,
-                   const Scalar& beta1,
-                   const Scalar& beta2,
-                   const Scalar& epsilon,
-                   bool lazy_mode,
-                   int64_t min_row_size_to_use_multithread,
-                   bool multi_precision,
-                   bool use_global_beta_pow,
-                   MetaTensor* param_out,
-                   MetaTensor* moment1_out,
-                   MetaTensor* moment2_out,
-                   MetaTensor* beta1_pow_out,
-                   MetaTensor* beta2_pow_out,
-                   MetaTensor* master_param_outs);
-
-void AdamwInferMeta(const MetaTensor& param,
-                    const MetaTensor& grad,
-                    const MetaTensor& learning_rate,
-                    const MetaTensor& moment1,
-                    const MetaTensor& moment2,
-                    const MetaTensor& beta1_pow,
-                    const MetaTensor& beta2_pow,
-                    paddle::optional<const MetaTensor&> master_param,
-                    paddle::optional<const MetaTensor&> skip_update,
-                    const Scalar& beta1,
-                    const Scalar& beta2,
-                    const Scalar& epsilon,
-                    float lr_ratio,
-                    float coeff,
-                    bool with_decay,
-                    bool lazy_mode,
-                    int64_t min_row_size_to_use_multithread,
-                    bool multi_precision,
-                    bool use_global_beta_pow,
-                    MetaTensor* param_out,
-                    MetaTensor* moment1_out,
-                    MetaTensor* moment2_out,
-                    MetaTensor* beta1_pow_out,
-                    MetaTensor* beta2_pow_out,
-                    MetaTensor* master_param_outs);
-
 void AddNInferMeta(const std::vector<MetaTensor*>& x,
                    MetaTensor* out,
                    MetaConfig config = MetaConfig());
